M3-1435: move tags pannel to a separate component
